% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{calc_correlation}
\alias{calc_correlation}
\title{Calculates the correlation coefficient between two variables}
\usage{
calc_correlation(x, y)
}
\arguments{
\item{explanatory}{The explanatory variable}

\item{response}{The response variable}
}
\value{
A htest object
}
\description{
Computes and returns a table with the pearson coefficient, as well as the
p-value, test statistic, and a confidence interval
}
